Data.Default

Description

This module defines a class for types with a default value. Instances are provided for (), Set, Map, Int, Integer, Float, Double, and many others (see below).

Documentation

class Default a where

A class for types with a default value.

Minimal complete definition

Nothing

Methods

def :: a

The default value for this type.

default def :: (Generic a, GDefault (Rep a)) => a